server [REVISE] all

+++ server/controllers/comment.js
@@ -12,7 +12,7 @@ module.exports = {
       comment: comment,
     });
 
-    res.status(200).json({ message: 'Comment Succesfully added' });
+    res.status(200).json({ message: '댓글을 추가했습니다.' });
   },
   get: async (req, res) => {
     // const result = await User.findOne({
@@ -25,8 +25,22 @@ module.exports = {
     // res.status(200)
   },
   put: async (req, res) => {
-    const { id } = req.params;
+    const { id } = req.params; // 게시물의 id
     const { commentId, comment } = req.body;
+
+    const comments = await db.sequelize.models.Comment.findOne({
+      where: {
+        id: commentId,
+        UserId: req.userId,
+        BoardId: id,
+      },
+    });
+
+    // console.log(comments);
+
+    if (!comments) {
+      return res.status(400).json({ message: '유저가 일치하지 않습니다' });
+    }
 
     await db.sequelize.models.Comment.update(
       {
@@ -41,25 +55,32 @@ module.exports = {
       },
     );
 
-    res.status(200).json({ message: 'Comment Succesfully modified' });
+    res.status(200).json({ message: '댓글을 수정 했습니다.' });
   },
   remove: async (req, res) => {
     const { id } = req.params;
-    const { commentId, comment } = req.body;
+    const { commentId } = req.body;
 
-    await db.sequelize.models.Comment.destroy(
-      {
-        comment: comment,
+    const comments = await db.sequelize.models.Comment.findOne({
+      where: {
+        id: commentId,
+        UserId: req.userId,
+        BoardId: id,
       },
-      {
-        where: {
-          id: commentId,
-          UserId: req.userId,
-          BoardId: id,
-        },
-      },
-    );
+    });
 
-    res.status(200).json({ message: 'Comment Succesfully deleted' });
+    if (!comments) {
+      return res.status(400).json({ message: '유저가 일치하지 않습니다' });
+    }
+
+    await db.sequelize.models.Comment.destroy({
+      where: {
+        id: commentId,
+        UserId: req.userId,
+        BoardId: id,
+      },
+    });
+
+    res.status(200).json({ message: '댓글을 삭제했습니다.' });
   },
 };
